backend/requirements.txt
asgiref==3.10.0
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
fonttools==4.60.1
gunicorn==23.0.0
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
